test: Move `Object`-related js_globals tests to submodule file
